Deleted old architucture
Preparing for new, simpler app's architecture
