+++ Calculator/Model/Operation/DivideOperation.cs
@@ -12,6 +12,7 @@ namespace Calculator.Model.Operation
 
         public override Number Execute()
         {
+            // TODO execute divide
             return _firstNumber.Divide(_secondNumber);
         }
     }
